refactor: extract interfaces from google and github use cases files for better modularity
