+++ src/context/auth/application/social/github/github.use-case.ts
@@ -13,6 +13,7 @@ import {
 import { AppConfigService } from 'src/global/services/app-config.service';
 import { GithubSocialLoginDTO } from './github.dto';
 import { AuthSharedService } from '../../shared/auth-shared.service';
+import { GitHubEmail, GitHubUser } from './github.types';
 
 @Injectable()
 export class GithubSocialLoginUseCase {
@@ -124,48 +125,4 @@ export class GithubSocialLoginUseCase {
 
     return accessToken;
   }
-}
-
-interface GitHubUser {
-  login: string;
-  id: number;
-  node_id: string;
-  avatar_url: string;
-  gravatar_id: string;
-  url: string;
-  html_url: string;
-  followers_url: string;
-  following_url: string;
-  gists_url: string;
-  starred_url: string;
-  subscriptions_url: string;
-  organizations_url: string;
-  repos_url: string;
-  events_url: string;
-  received_events_url: string;
-  type: string;
-  user_view_type: string;
-  site_admin: boolean;
-  name: string | null;
-  company: string | null;
-  blog: string;
-  location: string | null;
-  email: string | null;
-  hireable: boolean | null;
-  bio: string | null;
-  twitter_username: string | null;
-  notification_email: string | null;
-  public_repos: number;
-  public_gists: number;
-  followers: number;
-  following: number;
-  created_at: string;
-  updated_at: string;
-}
-
-interface GitHubEmail {
-  email: string;
-  primary: boolean;
-  verified: boolean;
-  visibility: 'public' | 'private' | null;
 }
