+++ src/context/auth/application/social/google/google.use-case.ts
@@ -12,31 +12,7 @@ import {
 } from '../../../domain';
 import { GoogleSocialLoginDTO } from './google.dto';
 import { AuthSharedService } from '../../shared/auth-shared.service';
-
-interface GoogleUserCredentials {
-  access_token: string;
-  expires_in: number;
-  refresh_token: string;
-  scope: string;
-  token_type: string;
-  id_token: string;
-}
-
-export interface GoogleUserInfo {
-  iss: string;
-  azp: string;
-  aud: string;
-  sub: string;
-  email: string;
-  email_verified: boolean;
-  at_hash: string;
-  name: string;
-  picture: string;
-  given_name: string;
-  family_name: string;
-  iat: number;
-  exp: number;
-}
+import { GoogleUserCredentials, GoogleUserInfo } from './google.types';
 
 @Injectable()
 export class GoogleSocialLoginUseCase {
